[Location] allow to del unused location only

+++ Wisky/Controllers/LocationController.cs
@@ -13,7 +13,7 @@ namespace DSS.Controllers
     public class LocationController : Controller
     {
         ILocationService locationService = DependencyUtils.Resolve<ILocationService>();
-        IMapper mapper = DependencyUtils.Resolve<IMapper>();        
+        IMapper mapper = DependencyUtils.Resolve<IMapper>();
         //GET: Location/Index
         public ActionResult Index()
         {
@@ -70,7 +70,7 @@ namespace DSS.Controllers
                         };
                         locationVMs.Add(b);
                     }
-                }               
+                }
                 return Json(new
                 {
                     locationIdDelete = locationVMs,
@@ -96,7 +96,7 @@ namespace DSS.Controllers
                     BrandId = item.BrandID,
                     District = item.District,
                     Province = item.Province,
-                    LocationId = item.LocationID,                
+                    LocationId = item.LocationID,
                 };
                 locationVMs.Add(b);
             }
@@ -140,7 +140,7 @@ namespace DSS.Controllers
             if (ModelState.IsValid)
             {
                 var location = new Data.Models.Entities.Location
-                {                   
+                {
                     BrandID = user.BrandID,
                     Province = model.Province,
                     District = model.District,
@@ -191,7 +191,11 @@ namespace DSS.Controllers
         public ActionResult Delete(int id)
         {
             var location = this.locationService.Get(id);
-            if (location != null)
+            IBoxService boxService = DependencyUtils.Resolve<IBoxService>();
+            IScreenService screenService = DependencyUtils.Resolve<IScreenService>();
+            var boxInLocation = boxService.Get(a => a.LocationID == id).FirstOrDefault();
+            var screenInLocation = screenService.Get(a => a.LocationID == id).FirstOrDefault();
+            if (location != null && boxInLocation == null && screenInLocation == null)
             {
                 this.locationService.Delete(location);
             }
